pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knomly"
version = "0.1.0"
description = "A Pipecat-inspired pipeline framework for building AI-powered voice and messaging applications"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Kuzushi Labs", email = "hello@kuzushilabs.com" }
]
maintainers = [
    { name = "Kuzushi Labs", email = "hello@kuzushilabs.com" }
]
keywords = [
    "ai",
    "pipeline",
    "voice",
    "messaging",
    "whatsapp",
    "llm",
    "speech-to-text",
    "framework",
    "pipecat",
    "agent",
    "tools",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Communications :: Chat",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Typing :: Typed",
]

# Core dependencies - minimal for the framework
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
# STT Providers
stt-gemini = ["google-generativeai>=0.3.0"]
stt-whisper = ["openai>=1.3.0"]
stt-deepgram = ["deepgram-sdk>=3.0.0"]

# LLM Providers
llm-openai = ["openai>=1.3.0"]
llm-anthropic = ["anthropic>=0.7.0"]
llm-gemini = ["google-generativeai>=0.3.0"]

# Chat Providers
chat-zulip = ["zulip>=0.9.0"]

# Transport Adapters
transport-twilio = ["twilio>=8.10.0"]
transport-telegram = ["python-telegram-bot>=20.0"]

# Audio Processing
audio = ["pydub>=0.25.0"]

# Database (for configuration storage)
mongodb = ["motor>=3.3.0"]

# Web Framework (for webhook handlers)
web = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
]

# All providers bundle
providers = [
    "knomly[stt-gemini,stt-whisper,llm-openai,llm-anthropic,chat-zulip]",
]

# All transports bundle
transports = [
    "knomly[transport-twilio]",
]

# Full installation
full = [
    "knomly[providers,transports,audio,mongodb,web]",
]

# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

# Documentation
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/kuzushi-labs/knomly"
Documentation = "https://github.com/kuzushi-labs/knomly#readme"
Repository = "https://github.com/kuzushi-labs/knomly"
Issues = "https://github.com/kuzushi-labs/knomly/issues"
Changelog = "https://github.com/kuzushi-labs/knomly/blob/main/CHANGELOG.md"

# =============================================================================
# Build Configuration
# =============================================================================

[tool.hatch.build.targets.wheel]
packages = ["knomly"]

[tool.hatch.build.targets.sdist]
include = [
    "/knomly",
    "/tests",
    "/examples",
    "/docs",
    "/configs",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]

# =============================================================================
# Tool Configuration
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["knomly"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["knomly", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip strict (Python 3.10+)
    "ARG001", # unused function argument (common in protocols)
    "ARG002", # unused method argument (common in processor interfaces)
    "B904",   # raise from (style preference, not always needed)
    "B024",   # abstract class without abstract methods (base classes)
    "B027",   # empty method in abstract class without decorator
    "RUF012", # mutable class attributes (common in dataclasses)
]

[tool.ruff.lint.isort]
known-first-party = ["knomly"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG001", "ARG002", "ARG005", "S101", "F841", "E402", "B017", "PTH123", "RUF013"]
"examples/*" = ["T201", "ARG002", "F841"]  # allow print, unused args and vars in examples
# Pydantic models need runtime imports, not TYPE_CHECKING
"knomly/adapters/schemas/*" = ["TC001", "TC003"]
"knomly/integrations/*/schemas.py" = ["TC001", "TC003"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
